# Conflicts:
#	agentManage/agentManage-common/src/main/java/com/ryx/credit/common/enumc/TabId.java

+++ agentManage/agentManageAdmin-dubbo/src/main/java/com/ryx/credit/dao/order/OReceiptProMapper.java
@@ -6,7 +6,7 @@ import com.ryx.credit.pojo.admin.order.OReceiptProExample;
 import java.util.List;
 
 public interface OReceiptProMapper {
-    int countByExample(OReceiptProExample example);
+    long countByExample(OReceiptProExample example);
 
     int deleteByExample(OReceiptProExample example);
